feat: View 인터페이스 구현
